Experimental changes to support Milvus 2.x
