+++ haystack/document_store/__init__.py
@@ -1,6 +1,6 @@
 from haystack.document_store.elasticsearch import ElasticsearchDocumentStore, OpenDistroElasticsearchDocumentStore, OpenSearchDocumentStore
 from haystack.document_store.faiss import FAISSDocumentStore
 from haystack.document_store.memory import InMemoryDocumentStore
-from haystack.document_store.milvus import MilvusDocumentStore
+from haystack.document_store.milvus2x import MilvusDocumentStore
 from haystack.document_store.sql import SQLDocumentStore
 from haystack.document_store.weaviate import WeaviateDocumentStore
